+++ record_web/lib/recorder/delegate/recorder_delegate.dart
@@ -1,6 +1,9 @@
+import 'dart:math';
+
 import 'package:flutter/foundation.dart';
 import 'package:record_platform_interface/record_platform_interface.dart';
 import 'package:record_web/js/js_interop/audio_context.dart';
+import 'package:record_web/js/js_interop/core.dart';
 
 typedef OnStateChanged = void Function(RecordState state);
 
@@ -23,7 +26,9 @@ abstract class RecorderDelegate {
 
   Future<String?> stop();
 
-  Future<MediaStream> initMediaStream(RecordConfig config) async {
+  Future<(MediaStream, RecordConfig)> initMediaStream(
+    RecordConfig config,
+  ) async {
     final constraints = MediaStreamConstraints(
       audio: config.device == null
           ? true
@@ -37,17 +42,21 @@ abstract class RecorderDelegate {
     );
     final audioTracks = mediaStream.getAudioTracks();
 
+    var adjustedConfig = config;
+
     for (var track in audioTracks) {
+      adjustedConfig = _adjustConfig(track, config);
+
       track.applyConstraints(MediaTrackConstraints(
-        autoGainControl: config.autoGain,
-        echoCancellation: config.echoCancel,
-        noiseSuppression: config.noiseSuppress,
-        sampleRate: config.sampleRate.toDouble(),
-        channelCount: config.numChannels.toDouble(),
+        autoGainControl: adjustedConfig.autoGain,
+        echoCancellation: adjustedConfig.echoCancel,
+        noiseSuppression: adjustedConfig.noiseSuppress,
+        sampleRate: adjustedConfig.sampleRate.toDouble(),
+        channelCount: adjustedConfig.numChannels.toDouble(),
       ));
     }
 
-    return mediaStream;
+    return (mediaStream, adjustedConfig);
   }
 
   Future<void> resetContext(
@@ -75,5 +84,38 @@ abstract class RecorderDelegate {
         debugPrint(e.toString());
       }
     }
+  }
+
+  RecordConfig _adjustConfig(MediaStreamTrack track, RecordConfig config) {
+    try {
+      final capabilities = track.getCapabilities();
+
+      return RecordConfig(
+        bitRate: config.bitRate,
+        device: config.device,
+        encoder: config.encoder,
+        autoGain: config.autoGain
+            ? capabilities.autoGainControl.any((e) => e == true)
+            : false,
+        echoCancel: config.echoCancel
+            ? capabilities.echoCancellation.any((e) => e == true)
+            : false,
+        noiseSuppress: config.noiseSuppress
+            ? capabilities.noiseSuppression.any((e) => e == true)
+            : false,
+        sampleRate: min(
+          max(config.sampleRate, capabilities.sampleRate.min),
+          capabilities.sampleRate.max,
+        ),
+        numChannels: min(
+          max(config.sampleRate, capabilities.channelCount.min),
+          capabilities.channelCount.max,
+        ),
+      );
+    } catch (error) {
+      debugPrint('getCapabilities error:\n$error');
+    }
+
+    return config;
   }
 }